+++ src/main/java/com/example/nutriplan_backend/controller/DatosNutricionalesController.java
@@ -2,12 +2,10 @@ package com.example.nutriplan_backend.controller;
 
 import com.example.nutriplan_backend.model.DatosNutricionales;
 import com.example.nutriplan_backend.service.DatosNutricionalesService;
-
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
+import org.springframework.http.ResponseEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
@@ -21,7 +19,7 @@ public class DatosNutricionalesController {
 
     //GET: Obtener por id
     @GetMapping("/{id}")
-    public ResponseEntity<DatosNutricionales> obtenerPorId(@PathVariable Long id){
+    public ResponseEntity<DatosNutricionales> obtenerPorId(@PathVariable Integer id){
         DatosNutricionales consulta = datosNutricionalesService.obtenerRegistrosPorUsuario(id);
         return ResponseEntity.ok(consulta);
     }
@@ -36,7 +34,7 @@ public class DatosNutricionalesController {
     //PUT: Actualizar un registro 
     @PutMapping("/{id}")
     public ResponseEntity <DatosNutricionales> actualizarRegistro(
-    @PathVariable Long id,    
+    @PathVariable Integer id,    
     @RequestBody DatosNutricionales dn){
         DatosNutricionales dnActualizado = datosNutricionalesService.actualizarDatos(id,dn);
         return ResponseEntity.ok(dnActualizado);
@@ -44,7 +42,7 @@ public class DatosNutricionalesController {
 
     //DELETE: Eliminar un registro 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> removerRegistro(@PathVariable Long id){
+    public ResponseEntity<Void> removerRegistro(@PathVariable Integer id){
         datosNutricionalesService.eliminarRegistro(id);
         return ResponseEntity.noContent().build(); 
     }
